First pass at a very preliminary Spark templating view factory.
It's largely based on the MVC one. I was unable to find an easy way
to get the kind of flexibility in view paths that we needed. The
MVC one seemed to rely too heavily on the controller pattern just..
everywhere. This is basically the MVC one stripped down. Most of
the functionality that I removed was the caching and the precompile
stuff, which we can get back later.

It still uses other parts of Spark.Web.Mvc2, like the SparkView
and all that, which I think will be useful in the beginning. So
I opted to do that instead of just build on top of the Spark core.